chore: refactor files and folders
