+++ src/auditLog/auditLog.controller.ts
@@ -1,7 +1,7 @@
 
 import { JwtAuthMiddleware } from "../auth/jwt/middleware/jwt-auth.middleware";
-import { Roles } from "../auth/roles.decorator";
-import { RolesGuard } from "../auth/roles.guard";
+import { Roles } from "../auth/guards/roles.decorator";
+import { RolesGuard } from "../auth/guards/roles.guard";
 import {
   Controller,
   DefaultValuePipe,
