+++ src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
-import { AuthController } from "./auth.controller";
-import { AuthService } from "./auth.service";
+import { AuthController } from "./controllers/auth.controller";
+import { AuthService } from "./services/auth.service";
 import { JwtService } from "./jwt/jwt.service";
 import {AuditLogModule} from "../auditLog/auditLog.module";
