+++ src/auth/guards/roles.guard.ts
@@ -5,7 +5,7 @@ import {
   Injectable,
 } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
-import { PrismaService } from "../prisma/prisma.service";
+import { PrismaService } from "../../prisma/prisma.service";
 import { ROLES_KEY } from "./roles.decorator";
 
 @Injectable()
