+++ src/auth/services/auth.service.ts
@@ -3,12 +3,12 @@ import {
   ForbiddenException,
   Injectable,
 } from "@nestjs/common";
-import { PrismaService } from "../prisma/prisma.service";
-import { AuthDto } from "./dto";
+import { PrismaService } from "../../prisma/prisma.service";
+import { AuthDto } from "../dto";
 import * as argon from "argon2";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
-import { JwtService } from "./jwt/jwt.service";
-import { AuditLogService } from "../auditLog/auditLog.service";
+import { JwtService } from "../jwt/jwt.service";
+import { AuditLogService } from "../../auditLog/auditLog.service";
 
 @Injectable()
 export class AuthService {
